feat: load protected mode kernel and exec

+++ lib.h
@@ -2,4 +2,4 @@ void flush();
 void print_str(const char *str);
 
 void seek(uint32 sectors);
-int16 read(uint8 *buffer, uint32 sectors);
+int16 read(uint8 *buffer, uint16 sectors);
